Allocate the cinfo/dinfo/jerr structures with C.malloc instead of Go.
This ought to avoid pitfalls related to passing Go pointers to C. I
still think there's a bug in the Go runtime causing the issue, but
everyone seems to advocate moving away from allocating C objects on
the Go heap. The image buffers and row arrays are still in Go land for
now; we'll see what needs to happen with those when the Go folks come up
with a concrete proposal.